% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projProp-package.R
\docType{package}
\name{projPropCCMS-package}
\alias{projPropCCMS}
\alias{projPropCCMS-package}
\title{projPropCCMS: Proposal Project Directory Generator adjusted to fit the needs of CCMS}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Create a project directory structure, along with typical files for that project. This allows projects to be quickly and easily created, as well as for them to be standardized. Designed specifically with scientists in mind (mainly bio-medical researchers, but likely applies to other fields) and adapted for CCMS.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rostools/prodigenr}
  \item \url{https://rostools.github.io/prodigenr/}
  \item \url{https://github.com/ccmsLIH/projPropCCMS}
  \item Report bugs at \url{https://github.com/rostools/prodigenr/issues}
}

}
\author{
\strong{Maintainer}: Luke Johnston \email{lwjohnst@gmail.com} (\href{https://orcid.org/0000-0003-4169-2616}{ORCID})

}
\keyword{internal}
